services/voice-pipeline/requirements.txt
# Pipecat core + WebSocket transport + Silero VAD
pipecat-ai[websocket,silero]>=0.0.100
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
python-json-logger>=2.0.0
prometheus-client>=0.20.0
